requirements.txt
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.0
flask-cors>=4.0.0
face-recognition>=1.3.0
opencv-python>=4.10.0.0
numpy>=1.26.0
Pillow>=10.0.0
python-dotenv>=1.0.1
pytz>=2024.1
